+++ src/main/scala/grizzled/math/stats.scala
@@ -24,22 +24,22 @@ object stats {
     * arithmetic mean, see
     * [[http://www.math.toronto.edu/mathnet/questionCorner/geomean.html]].
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the geometric mean
     */
-  def geometricMean[T](items: T*)(implicit n: Numeric[T]): Double = {
-    val itemList = items.toList
-    val len = itemList.length
-    require (len > 0)
+  def geometricMean[T: Numeric](item: T, items: T*): Double = {
+    val itemList = (item +: items).toList
+    val n = implicitly[Numeric[T]]
     require (! itemList.exists(n.toDouble(_) <= 0))
 
-    len match {
-      case 1 =>
-        n.toDouble(itemList.head)
+    items match {
+      case seq if seq.isEmpty =>
+        n.toDouble(item)
 
       case _ =>
-        val recip = 1.0 / len.toDouble
+        val recip = 1.0 / itemList.length.toDouble
         (1.0 /: itemList)((a, b) => a * math.pow(n.toDouble(b), recip))
     }
   }
@@ -47,65 +47,72 @@ object stats {
   /** Calculates the harmonic mean of the values of the passed-in
     * numbers, namely: `n / (1/x^1^ + 1/x^2^ + ... + 1/x^n^).`
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the harmonic mean
     */
-  def harmonicMean[T](items: T*)(implicit n: Numeric[T]): Double = {
-    val itemList = items.toList
-    val len = itemList.length
-    require (len > 0)
+  def harmonicMean[T: Numeric](item: T, items: T*): Double = {
+    val n = implicitly[Numeric[T]]
 
-    len match {
-      case 1 => n.toDouble(itemList.head)
-      case _ => len / (0.0 /: itemList)((a, b) => a + (1.0 / n.toDouble(b)))
+    items.toList match {
+      case Nil => n.toDouble(item)
+      case _ =>
+        val allItems = item +: items
+        allItems.length / (0.0 /: allItems)((a, b) => a + (1.0 / n.toDouble(b)))
     }
   }
 
   /** Calculates the arithmetic mean of the values of the passed-in
     * numbers.
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the arithmetic mean
     */
-  def arithmeticMean[T](items: T*)(implicit n: Numeric[T]): Double = {
-    val itemList = items.toList
-    val len = itemList.length
-    require (len > 0)
+  def arithmeticMean[T: Numeric](item: T, items: T*): Double = {
+    val n = implicitly[Numeric[T]]
 
-    len match {
-      case 1 => n.toDouble(itemList.head)
-      case _ => (0.0 /: itemList)((a, b) => a + n.toDouble(b)) / len
+    items.toList match {
+      case Nil => n.toDouble(item)
+      case _ =>
+        val allItems = item +: items
+        (0.0 /: allItems)((a, b) => a + n.toDouble(b)) / allItems.length
     }
   }
 
   /** Synonym for `arithmeticMean`.
     *
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
+    *
+    * @return the arithmetic mean
+    *
     * @see arithmeticMean
     */
-  def mean[T](items: T*)(implicit n: Numeric[T]): Double =
-    arithmeticMean(items.toList: _*)
+  def mean[T: Numeric](item: T, items: T*): Double =
+    arithmeticMean(item, items: _*)
 
 
   /** Calculates the median of the values of the passed-in numbers.
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the median
     */
-  def median[T](items: T*)(implicit n: Numeric[T]): Double = {
-    val itemList = items.toList
-    val len = itemList.length
-    require (len > 0)
+  def median[T: Numeric](item: T, items: T*): Double = {
+    val n = implicitly[Numeric[T]]
 
-    if (len == 1)
-      n.toDouble(itemList.head)
+    if (items.length == 0)
+      n.toDouble(item)
 
     else {
-      val sorted = itemList sortWith (n.compare(_, _) < 0)
+      val allItems = item +: items
+      val sorted = allItems sortWith (n.compare(_, _) < 0)
       val midpoint = sorted.length / 2
-      len % 2 match {
+      allItems.length % 2 match {
         case 0 => mean(n.toDouble(sorted(midpoint)),
                        n.toDouble(sorted(midpoint - 1)))
         case 1 => n.toDouble(sorted(midpoint))
@@ -117,29 +124,28 @@ object stats {
     * passed-in numbers. If there are multiple common values, they're all
     * returned.
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return list of modal values
     */
-  def mode[T](items: T*)(implicit n: Numeric[T]): List[T] = {
-    val itemList = items.toList
-    val len = itemList.length
-    require (len > 0)
-
-    if (len == 1)
-      itemList take 1
+  def mode[T: Numeric](item: T, items: T*): List[T] = {
+    if (items.length == 1)
+      List(item)
 
     else {
        // Count the occurrences of each value. This is a reduceByKey operation.
-      val m = itemList.map { n => n -> 1 }
-                      .groupBy(_._1)
-                      .map { case (n, counts) => n -> counts.map(_._2).sum }
+      val allItems = item +: items
+      val m = allItems
+        .map(_ -> 1)
+        .groupBy(_._1)
+        .map { case (n, counts) => n -> counts.map(_._2).sum }
 
       // Find the maximum count.
       val max = m.values.max
 
       // Extract the keys with values that match
-      m.filter { case (k, v) => v == max }.keys.toList
+      m.filter { case (_, v) => v == max }.keys.toList
     }
   }
 
@@ -157,12 +163,15 @@ object stats {
     * - [[http://en.wikipedia.org/wiki/Variance#Population_variance_and_sample_variance]]
     * - [[http://www.quickmba.com/stats/standard-deviation/]]
     *
-    * @param items  the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the variance
     */
-  def populationVariance[T](items: T*)(implicit n: Numeric[T]): Double =
-    calculateVariance(items.length, items.toList)
+  def populationVariance[T: Numeric](item: T, items: T*): Double = {
+    val allItems = item +: items
+    calculateVariance(allItems.length, allItems.toList)
+  }
 
   /** Calculate the unbiased sample variance of the finite sample defined
     * by the `items` arguments. The sample variance is defined as:
@@ -178,12 +187,15 @@ object stats {
     * - [[http://en.wikipedia.org/wiki/Variance#Population_variance_and_sample_variance]]
     * - [[http://www.quickmba.com/stats/standard-deviation/]]
     *
-    * @param items  the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the variance
     */
-  def sampleVariance[T](items: T*)(implicit n: Numeric[T]): Double =
-    calculateVariance(items.length - 1, items.toList)
+  def sampleVariance[T: Numeric](item: T, items: T*): Double = {
+    val allItems = item +: items
+    calculateVariance(allItems.length -1, allItems.toList)
+  }
 
   /** Calculate the population standard deviation of the specified values.
     * The population standard deviation is merely the square root of the
@@ -193,19 +205,23 @@ object stats {
     * java.lang.Math.sqrt(populationVariance(items))
     * }}}
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the standard deviation
     */
-  def populationStandardDeviation[T](items: T*)(implicit n: Numeric[T]): Double =
-    java.lang.Math.sqrt(populationVariance(items.toList: _*))
+  def populationStandardDeviation[T: Numeric](item: T, items: T*): Double =
+    java.lang.Math.sqrt(populationVariance(item, items: _*))
 
   /** Shorter synonym for `populationStandardDeviation`.
     *
-    * @see populationStandardDeviation
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
+    *
+    * @return the standard deviation
     */
-  def popStdDev[T](items: T*)(implicit n: Numeric[T]): Double =
-    java.lang.Math.sqrt(populationVariance(items.toList: _*))
+  def popStdDev[T: Numeric](item: T, items: T*): Double =
+    java.lang.Math.sqrt(populationVariance(item, items: _*))
 
   /** Calculate the sample standard deviation of the specified values.
     * The sample standard deviation is merely the square root of the
@@ -215,52 +231,64 @@ object stats {
     * java.lang.Math.sqrt(sampleVariance(items))
     * }}}
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
-    * @return the standard deviation
+    * @return the sample standard deviation
     */
-  def sampleStandardDeviation[T](items: T*)(implicit n: Numeric[T]): Double =
-    java.lang.Math.sqrt(sampleVariance(items.toList: _*))
+  def sampleStandardDeviation[T: Numeric](item: T, items: T*): Double =
+    java.lang.Math.sqrt(sampleVariance(item, items: _*))
 
   /** Shorter synonym for `sampleStandardDeviation`.
     *
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
+    *
+    * @return the sample standard deviation
+    *
     * @see populationStandardDeviation
     */
-  def sampleStdDev[T](items: T*)(implicit n: Numeric[T]): Double =
-    java.lang.Math.sqrt(sampleVariance(items.toList: _*))
+  def sampleStdDev[T: Numeric](item: T, items: T*): Double =
+    java.lang.Math.sqrt(sampleVariance(item, items: _*))
 
 
   /** Calculate the range of a data set. This function does a single
     * linear pass over the data set.
     *
-    * @param items the numbers on which to operate
+    * @param item  the first number on which to operate
+    * @param items the remaining numbers on which to operate
     *
     * @return the range
     */
-  def range[T](items: T*)(implicit n: Numeric[T]): T = {
+  def range[T: Numeric](item: T, items: T*): T = {
+    val n = implicitly[Numeric[T]]
     items.length match {
-      case 1 => n.minus(items(0), items(0))
+      case 0 => n.minus(item, item)
       case _ =>
         // Fold left, generating a (min, max) tuple along the way.
+        val allItems = item +: items
         val (min, max) =
-          ((n.fromInt(Int.MaxValue), n.fromInt(0)) /: items)((tuple, i) =>
+          ((n.fromInt(Int.MaxValue), n.fromInt(0)) /: allItems)((tuple, i) =>
            (n.min(tuple._1, i), n.max(tuple._1, i)))
         n.minus(max, min)
     }
   }
 
-  private def calculateVariance[T](denominator: Int, items: List[T])
-                                  (implicit n: Numeric[T]): Double = {
-
-    def sumOfSquares(dList: List[Double]): Double =
+  private def calculateVariance[T: Numeric](denominator: Int,
+                                            items: Seq[T]): Double = {
+    def sumOfSquares(dList: Seq[Double]): Double =
       (0.0 /: dList) ((sum, d) => sum + (d * d))
 
-    val itemList = items.toList
-    val len = itemList.length
-    require (len > 1)
+    require (items.length > 1)
+    val n = implicitly[Numeric[T]]
 
-    val mn = mean(itemList: _*)
-    val deviations = itemList map (n.toDouble(_) - mn)
-    sumOfSquares(deviations) / denominator.toDouble
+    items match {
+      case seq if seq.isEmpty =>
+        0
+      case Seq(head, tail @ _*) =>
+        val mn = mean(head, tail: _*)
+        val deviations = items map (n.toDouble(_) - mn)
+        sumOfSquares(deviations) / denominator.toDouble
+    }
   }
 }