+++ src/test/scala/grizzled/math/MathUtilSpec.scala
@@ -9,38 +9,38 @@ class MathUtilSpec extends BaseSpec {
   "max" should "work on floats" in {
     val nums = (1 to 100).map(_ => Random.nextFloat * 1000)
     val biggest = nums.sortWith { _ > _ }.head
-    maximum(nums: _*) shouldBe biggest
+    maximum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on integers" in {
     val nums = (1 to 100).map(_ => Random.nextInt(1000))
     val biggest = nums.sortWith { _ > _ }.head
-    maximum(nums: _*) shouldBe biggest
+    maximum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on bytes" in {
     val bytes = new Array[Byte](1000)
     Random.nextBytes(bytes)
     val biggest = bytes.sortWith { _ > _ }.head
-    maximum(bytes: _*) shouldBe biggest
+    maximum(bytes.head, bytes.tail: _*) shouldBe biggest
   }
 
   it should "work on doubles" in {
     val nums = (1 to 100).map(_ => Random.nextDouble * 100000000)
     val biggest = nums.sortWith { _ > _ }.head
-    maximum(nums: _*) shouldBe biggest
+    maximum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on BigInts" in {
     val nums = (1 to 100).map(_ => BigInt(Random.nextLong))
     val biggest = nums.sortWith { _ > _ }.head
-    maximum(nums: _*) shouldBe biggest
+    maximum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on BigDecimals" in {
     val nums = (1 to 100).map(_ => BigDecimal(Random.nextDouble * 100000000))
     val biggest = nums.sortWith { _ > _ }.head
-    maximum(nums: _*) shouldBe biggest
+    maximum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on Strings" in {
@@ -51,44 +51,44 @@ class MathUtilSpec extends BaseSpec {
       "wertyuijkl"
     )
     val biggest = strings.sortWith { _ > _ }.head
-    maximum(strings: _*) shouldBe biggest
+    maximum(strings.head, strings.tail: _*) shouldBe biggest
   }
 
   "min" should "work on floats" in {
     val nums = (1 to 100).map(_ => Random.nextFloat * 1000)
     val biggest = nums.sortWith { _ < _ }.head
-    minimum(nums: _*) shouldBe biggest
+    minimum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on integers" in {
     val nums = (1 to 100).map(_ => Random.nextInt(1000))
     val biggest = nums.sortWith { _ < _ }.head
-    minimum(nums: _*) shouldBe biggest
+    minimum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on bytes" in {
     val bytes = new Array[Byte](1000)
     Random.nextBytes(bytes)
     val biggest = bytes.sortWith { _ < _ }.head
-    minimum(bytes: _*) shouldBe biggest
+    minimum(bytes.head, bytes.tail: _*) shouldBe biggest
   }
 
   it should "work on doubles" in {
     val nums = (1 to 100).map(_ => Random.nextDouble * 100000000)
     val biggest = nums.sortWith { _ < _ }.head
-    minimum(nums: _*) shouldBe biggest
+    minimum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on BigInts" in {
     val nums = (1 to 100).map(_ => BigInt(Random.nextLong))
     val biggest = nums.sortWith { _ < _ }.head
-    minimum(nums: _*) shouldBe biggest
+    minimum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on BigDecimals" in {
     val nums = (1 to 100).map(_ => BigDecimal(Random.nextDouble * 100000000))
     val biggest = nums.sortWith { _ < _ }.head
-    minimum(nums: _*) shouldBe biggest
+    minimum(nums.head, nums.tail: _*) shouldBe biggest
   }
 
   it should "work on Strings" in {
@@ -99,6 +99,6 @@ class MathUtilSpec extends BaseSpec {
       "wertyuijkl"
     )
     val biggest = strings.sortWith { _ < _ }.head
-    minimum(strings: _*) shouldBe biggest
+    minimum(strings.head, strings.tail: _*) shouldBe biggest
   }
 }